Refactor service model with guards

+++ src/models/service/service.service.ts
@@ -1,5 +1,8 @@
 import {
     BadRequestException,
+    ForbiddenException,
+    forwardRef,
+    Inject,
     Injectable,
     NotFoundException,
 } from '@nestjs/common';
@@ -10,9 +13,13 @@ import { ConfigEnum } from 'common';
 import { WithPaginationResponse } from 'common/types';
 import { getPagPayload } from 'common/utils';
 import { Company } from 'models/company/entities/company.entity';
+import { EmployeeService } from 'models/employee/employee.service';
+import { ActiveUserData } from 'models/iam/decorators';
+import { checkIsAdmin } from 'models/iam/utils/checkIsAdmin';
 
 import { CreateServiceDto } from './dto/create-service.dto';
 import { FindAllServicesDto } from './dto/find-all-services.dto';
+import { RemoveServiceDto } from './dto/remove-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
 import { Service } from './entities/service.entity';
 
@@ -23,10 +30,26 @@ export class ServiceService {
         private readonly serviceRepository: Repository<Service>,
         @InjectRepository(Company, ConfigEnum.DB_CONNECTION_NAME)
         private readonly companyRepository: Repository<Company>,
+        @Inject(forwardRef(() => EmployeeService))
+        private readonly employeeService: EmployeeService,
     ) {}
 
-    async create(dto: CreateServiceDto): Promise<Service> {
+    async create(
+        user: ActiveUserData,
+        dto: CreateServiceDto,
+    ): Promise<Service> {
         const { name, company_uuid } = dto;
+
+        const isAdmin = checkIsAdmin(user);
+        const isCompanyEmployee = this.employeeService.checkIsCompanyEmployee(
+            user,
+            company_uuid,
+            user.employee_id,
+        );
+
+        if (!isAdmin && !isCompanyEmployee) {
+            throw new ForbiddenException();
+        }
 
         const sameNameService = await this.serviceRepository.findOneBy({
             name,
@@ -50,10 +73,22 @@ export class ServiceService {
     }
 
     async findAll(
+        user: ActiveUserData,
         dto: FindAllServicesDto,
     ): Promise<WithPaginationResponse<Service>> {
         const { employee_id, company_uuid } = dto;
         const { offset, payload } = getPagPayload(dto);
+
+        const isAdmin = checkIsAdmin(user);
+        const isCompanyEmployee = this.employeeService.checkIsCompanyEmployee(
+            user,
+            company_uuid,
+            user.employee_id,
+        );
+
+        if (!isAdmin && !isCompanyEmployee) {
+            throw new ForbiddenException();
+        }
 
         const [items, total] = await this.serviceRepository.findAndCount({
             ...payload,
@@ -85,13 +120,35 @@ export class ServiceService {
         return foundService;
     }
 
-    async update(uuid: string, dto: UpdateServiceDto) {
+    async update(user: ActiveUserData, uuid: string, dto: UpdateServiceDto) {
+        const { company_uuid } = dto;
+        const isAdmin = checkIsAdmin(user);
+        const isCompanyEmployee = this.employeeService.checkIsCompanyEmployee(
+            user,
+            company_uuid,
+            user.employee_id,
+        );
+
+        if (!isAdmin && !isCompanyEmployee) {
+            throw new ForbiddenException();
+        }
+
         await this.serviceRepository.update(uuid, dto);
 
         return this.findOne(uuid);
     }
 
-    remove(uuid: string) {
+    remove(user: ActiveUserData, uuid: string, dto: RemoveServiceDto) {
+        const isAdmin = checkIsAdmin(user);
+        const isCompanyOwner = this.employeeService.checkIsCompanyOwner(
+            user,
+            dto.company_uuid,
+        );
+
+        if (!isAdmin && !isCompanyOwner) {
+            throw new ForbiddenException();
+        }
+
         return this.serviceRepository.delete(uuid);
     }
 }
